+++ cpp/src/db/meta/MetaFactory.h
@@ -22,17 +22,17 @@
 
 #include <string>
 
-namespace zilliz {
 namespace milvus {
 namespace engine {
 
 class MetaFactory {
  public:
-    static DBMetaOptions BuildOption(const std::string &path = "");
+    static DBMetaOptions
+    BuildOption(const std::string& path = "");
 
-    static meta::MetaPtr Build(const DBMetaOptions &metaOptions, const int &mode);
+    static meta::MetaPtr
+    Build(const DBMetaOptions& metaOptions, const int& mode);
 };
 
-} // namespace engine
-} // namespace milvus
-} // namespace zilliz
+}  // namespace engine
+}  // namespace milvus