+++ cpp/src/scheduler/Utils.h
@@ -15,10 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
-
 #include <cstdint>
 
-namespace zilliz {
 namespace milvus {
 namespace scheduler {
 
@@ -28,6 +26,5 @@ get_current_timestamp();
 uint64_t
 get_num_gpu();
 
-} // namespace scheduler
-} // namespace milvus
-} // namespace zilliz
+}  // namespace scheduler
+}  // namespace milvus