+++ cpp/src/scheduler/resource/Resource.h
@@ -17,25 +17,24 @@
 
 #pragma once
 
-#include <string>
-#include <vector>
-#include <memory>
-#include <utility>
-#include <thread>
-#include <functional>
 #include <condition_variable>
+#include <functional>
+#include <memory>
+#include <string>
+#include <thread>
+#include <utility>
+#include <vector>
 
-#include "../event/Event.h"
-#include "../event/StartUpEvent.h"
-#include "../event/LoadCompletedEvent.h"
-#include "../event/FinishTaskEvent.h"
-#include "../event/TaskTableUpdatedEvent.h"
 #include "../TaskTable.h"
+#include "../event/Event.h"
+#include "../event/FinishTaskEvent.h"
+#include "../event/LoadCompletedEvent.h"
+#include "../event/StartUpEvent.h"
+#include "../event/TaskTableUpdatedEvent.h"
 #include "../task/Task.h"
 #include "Connection.h"
 #include "Node.h"
 
-namespace zilliz {
 namespace milvus {
 namespace scheduler {
 
@@ -99,7 +98,7 @@ class Resource : public Node, public std::enable_shared_from_this<Resource> {
         return device_id_;
     }
 
-    TaskTable &
+    TaskTable&
     task_table() {
         return task_table_;
     }
@@ -115,11 +114,11 @@ class Resource : public Node, public std::enable_shared_from_this<Resource> {
         return enable_executor_;
     }
 
-    // TODO: const
+    // TODO(wxyu): const
     uint64_t
     NumOfTaskToExec();
 
-    // TODO: need double ?
+    // TODO(wxyu): need double ?
     inline uint64_t
     TaskAvgCost() const {
         return total_cost_ / total_task_;
@@ -130,14 +129,11 @@ class Resource : public Node, public std::enable_shared_from_this<Resource> {
         return total_task_;
     }
 
-    friend std::ostream &operator<<(std::ostream &out, const Resource &resource);
+    friend std::ostream&
+    operator<<(std::ostream& out, const Resource& resource);
 
  protected:
-    Resource(std::string name,
-             ResourceType type,
-             uint64_t device_id,
-             bool enable_loader,
-             bool enable_executor);
+    Resource(std::string name, ResourceType type, uint64_t device_id, bool enable_loader, bool enable_executor);
 
     /*
      * Implementation by inherit class;
@@ -212,6 +208,5 @@ class Resource : public Node, public std::enable_shared_from_this<Resource> {
 using ResourcePtr = std::shared_ptr<Resource>;
 using ResourceWPtr = std::weak_ptr<Resource>;
 
-} // namespace scheduler
-} // namespace milvus
-} // namespace zilliz
+}  // namespace scheduler
+}  // namespace milvus