+++ cpp/src/scheduler/tasklabel/TaskLabel.h
@@ -19,14 +19,13 @@
 
 #include <memory>
 
-namespace zilliz {
 namespace milvus {
 namespace scheduler {
 
 enum class TaskLabelType {
-    DEFAULT, // means can be executed in any resource
-    SPECIFIED_RESOURCE, // means must executing in special resource
-    BROADCAST, // means all enable-executor resource must execute task
+    DEFAULT,             // means can be executed in any resource
+    SPECIFIED_RESOURCE,  // means must executing in special resource
+    BROADCAST,           // means all enable-executor resource must execute task
 };
 
 class TaskLabel {
@@ -46,6 +45,5 @@ class TaskLabel {
 
 using TaskLabelPtr = std::shared_ptr<TaskLabel>;
 
-} // namespace scheduler
-} // namespace milvus
-} // namespace zilliz
+}  // namespace scheduler
+}  // namespace milvus